# docs/swagger.yaml
definitions:
...
  shared-modules_entities.AutoYieldEnableForm:
    properties:
      cardId:
        example: "0"
        type: string
      enable:
        type: boolean
    required:
    - cardId
    - enable
    type: object
  shared-modules_entities.AutoYieldHistoryDataVO:
    properties:
      interest:
        type: string
      principalAmount:
        type: string
      timestamp:
        example: "0"
        type: string
    type: object
  shared-modules_entities.AutoYieldHistoryForm:
    properties:
      currency:
        type: string
      period:
        maximum: 1000
        type: integer
    required:
    - currency
    - period
    type: object
  shared-modules_entities.AutoYieldHistoryVO:
    properties:
      records:
        items:
          $ref: '#/definitions/shared-modules_entities.AutoYieldHistoryDataVO'
        type: array
    type: object
  shared-modules_entities.AutoYieldInfoForm:
    properties:
      currency:
        type: string
    required:
    - currency
    type: object
  shared-modules_entities.AutoYieldInfoVO:
    properties:
      annualYieldRate:
        type: number
      cardId:
        example: "0"
        type: string
      dailyInterest:
        type: string
      earningStatus:
        type: string
      oneMonthAccumulatedInterest:
        type: string
      principalAmount:
        type: string
      threeMonthsAccumulatedInterest:
        type: string
      threshold:
        type: number
      thresholdCurrency:
        type: string
      twoMonthsAccumulatedInterest:
        type: string
    type: object
  shared-modules_entities.AutoYieldInterestList:
    properties:
      isEarning:
        type: boolean
      records:
        items:
          $ref: '#/definitions/shared-modules_entities.AutoYieldInterestVO'
        type: array
    type: object
  shared-modules_entities.AutoYieldInterestVO:
    properties:
      annualYieldRate:
        type: number
      currency:
        type: string
      dailyInterest:
        type: string
      earningStatus:
        type: string
      principalAmount:
        type: string
    type: object
...
  shared-modules_entities.GetTransactionRecordForm:
    properties:
      cardId:
        example: "0"
        type: string
      orderNO:
        type: string
      side:
        type: string
    type: object
...
  shared-modules_entities.PageAutoYieldTransactionRecordsForm:
    properties:
      currency:
        type: string
      current:
        minimum: 1
        type: integer
      pageSize:
        maximum: 3000
        type: integer
      types:
        items:
          type: string
        type: array
    required:
    - current
    - pageSize
    type: object
...
  shared-modules_entities.PageTransactionRecordsForm:
    properties:
      cardId:
        example: "0"
        type: string
      category:
        type: string
      categoryId:
        example: "0"
        type: string
      current:
        minimum: 1
        type: integer
      pageSize:
        maximum: 3000
        type: integer
      types:
        items:
          type: string
        type: array
    required:
    - current
    - pageSize
    type: object
...
info:
  contact: {}
paths:
...
  /web/financial/autoYield/enable:
    post:
      parameters:
      - description: body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/shared-modules_entities.AutoYieldEnableForm'
      - description: User token
        in: header
        name: X-Token
        required: true
        type: string
      - description: accept language
        in: header
        name: Accept-Language
        type: string
      responses: {}
      tags:
      - web/financial
  /web/financial/autoYield/history:
    post:
      parameters:
      - description: body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/shared-modules_entities.AutoYieldHistoryForm'
      - description: User token
        in: header
        name: X-Token
        required: true
        type: string
      - description: accept language
        in: header
        name: Accept-Language
        type: string
      responses:
        "0":
          description: data
          schema:
            $ref: '#/definitions/shared-modules_entities.AutoYieldHistoryVO'
      tags:
      - web/financial
  /web/financial/autoYield/info:
    post:
      parameters:
      - description: body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/shared-modules_entities.AutoYieldInfoForm'
      - description: User token
        in: header
        name: X-Token
        required: true
        type: string
      - description: accept language
        in: header
        name: Accept-Language
        type: string
      responses:
        "0":
          description: data
          schema:
            $ref: '#/definitions/shared-modules_entities.AutoYieldInfoVO'
      tags:
      - web/financial
  /web/financial/autoYield/list:
    post:
      parameters:
      - description: User token
        in: header
        name: X-Token
        required: true
        type: string
      - description: accept language
        in: header
        name: Accept-Language
        type: string
      responses:
        "0":
          description: data
          schema:
            $ref: '#/definitions/shared-modules_entities.AutoYieldInterestList'
      tags:
      - web/financial
...
  /web/order/transactions/autoYield/page:
    post:
      consumes:
      - application/json
      description: Page List transaction records based on provided filters. <br> walletID和cardID和categoryID擇一放入
      parameters:
      - description: body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/shared-modules_entities.PageAutoYieldTransactionRecordsForm'
      - description: User token
        in: header
        name: X-Token
        required: true
        type: string
      - description: accept language
        in: header
        name: Accept-Language
        type: string
      - description: Extend
        in: header
        name: X-Extend
        type: string
      - description: Convert
        in: header
        name: X-Convert
        type: string
      produces:
      - application/json
      responses: {}
      summary: Page List Transaction Records
      tags:
      - web/order
  /web/order/transactions/get:
    post:
      consumes:
      - application/json
      description: Get transaction record based on provided filters.
      parameters:
      - description: body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/shared-modules_entities.GetTransactionRecordForm'
      - description: User token
        in: header
        name: X-Token
        required: true
        type: string
      - description: accept language
        in: header
        name: Accept-Language
        type: string
      produces:
      - application/json
      responses: {}
      summary: Get Transaction Record
      tags:
      - web/order
  /web/order/transactions/page:
    post:
      consumes:
      - application/json
      description: Page List transaction records based on provided filters. <br> walletID和cardID和categoryID擇一放入
      parameters:
      - description: body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/shared-modules_entities.PageTransactionRecordsForm'
      - description: User token
        in: header
        name: X-Token
        required: true
        type: string
      - description: accept language
        in: header
        name: Accept-Language
        type: string
      - description: Extend
        in: header
        name: X-Extend
        type: string
      - description: Convert
        in: header
        name: X-Convert
        type: string
      produces:
      - application/json
      responses: {}
      summary: Page List Transaction Records
      tags:
      - web/order
...
swagger: "2.0"